Rename equilibrium_result.py to light_dataset.py
